Selenium Test Improved.
